feat(checker.c): Bonus makefil규칙 설정 및 checker 구현 list sa 함수 수정

+++ libft/ft_strlen.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned long	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
 	int	i;
